Israel/Palestine War
This analysis evaluate the ongoing war between Israel and Palestine. It indicated the citizens and the type of war experienced. It look at the regions and type of war common in those regions. Hostilities by gender/citizens and quarter of Death experienced. This shows a high rate of Palestine being a victim of the ongoing war. 

Dataset Source: Kaggle