Global Terrorism Attack
The visualization shows the activities of terrorism on a global state. From the KPI’s identified, the following were identified: 
	Total Number of the attacks 
	Number of children that died 
	Number of Adults that died 
	Number of older people that died 
	Total number of terrorists that was killed 
The analysis showed that Bombing/Explosion are the attack method used mostly by the terrorists. The analysis was further visualized by the attacks and death resulting from those attacks in the last 8 years. It was noted that 2014 experienced the highest terrorists attacks and death. Across all the years, 2014 recalled the highest terrorism death and children were mostly the victim of death. On the continents where terrorists are mostly killed, it was observed that the killing of the terrorists took place mostly in Asia continent. 
Recommendations
There is need for the government and other stakeholders to reduce the production of explosive materials so as to reduce terrorists access to these materials. Moreover, there is need for stakeholders to under the course of time and see to the reasons why terrorist attack was high across years and put in measures to further reduce its occurrence. Children are mostly affected by terrorism, there is need for the government to consider the evacuation of children more in the incidents of terrorist attack.